requirements.txt
Flask==3.0.3
google-generativeai==0.8.3
google-api-python-client==2.151.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.65.0
grpcio==1.67.1
requests==2.32.3
Werkzeug==3.1.3
Jinja2==3.1.4
tqdm==4.67.0
